src/swarm_registry/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-instrumentation-fastapi==0.54b1
azure-monitor-opentelemetry-exporter==1.0.0b35
